+++ lib/hi_energy_ai/resources/clicks.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class Clicks < Resource
+      def list(start_date:, end_date:, **params)
+        get("/clicks", params: params.merge(start_date: start_date, end_date: end_date))
+      end
+    end
+  end
+end