+++ lib/hi_energy_ai/resources/networks.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class Networks < Resource
+      def list(**params)
+        get("/networks", params: params)
+      end
+
+      def find(id, **params)
+        get("/networks/#{id}", params: params)
+      end
+    end
+  end
+end