+++ lib/hi_energy_ai/resources/schema.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class Schema < Resource
+      def fetch(**params)
+        get("/schema", params: params)
+      end
+    end
+  end
+end