+++ lib/hi_energy_ai/resources/status_changes.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class StatusChanges < Resource
+      def list(**params)
+        get("/status_changes", params: params)
+      end
+    end
+  end
+end