+++ lib/hi_energy_ai/resources/tags.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class Tags < Resource
+      def list(**params)
+        get("/tags", params: params)
+      end
+
+      alias search list
+
+      def advertisers(id, **params)
+        get("/tags/#{id}/advertisers", params: params)
+      end
+    end
+  end
+end