+++ lib/hi_energy_ai/resources/verticals.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HiEnergyAi
+  module Resources
+    class Verticals < Resource
+      def list(**params)
+        get("/verticals", params: params)
+      end
+    end
+  end
+end